+++ hindsight-api-slim/hindsight_api/extensions/__init__.py
@@ -13,15 +13,16 @@ Example:
 
 Extensions receive an ExtensionContext that provides a controlled API for interacting
 with the system (e.g., running migrations for tenant schemas).
+
+This package exports the extension *interfaces* only. Concrete implementations —
+including the ones bundled under ``hindsight_api.extensions.builtin`` — are
+imported by path at load time, so importing this package never pulls in an
+implementation's dependencies. Extensions distributed outside the server live in
+``hindsight-extensions/`` in the repository.
 """
 
 from hindsight_api.extensions.bank_tables import BankScopedTable
 from hindsight_api.extensions.base import Extension
-from hindsight_api.extensions.builtin import (
-    ApiKeyTenantExtension,
-    MemoryDefenseRegexExtension,
-    SupabaseTenantExtension,
-)
 from hindsight_api.extensions.context import DefaultExtensionContext, ExtensionContext
 from hindsight_api.extensions.http import HttpExtension
 from hindsight_api.extensions.loader import load_extension
@@ -120,8 +121,6 @@ __all__ = [
     "MentalModelRefreshContext",
     "MentalModelRefreshResult",
     # Tenant/Auth
-    "ApiKeyTenantExtension",
-    "SupabaseTenantExtension",
     "AuthenticationError",
     "RequestContext",
     "Tenant",
@@ -132,7 +131,6 @@ __all__ = [
     "DefenseDecision",
     "DefensePolicy",
     "MemoryDefenseExtension",
-    "MemoryDefenseRegexExtension",
     "PolicyRule",
     "apply_redaction",
     "parse_policy",
