+++ hindsight-extensions/supabase-tenant/pyproject.toml
@@ -1,0 +1,55 @@
+[build-system]
+requires = ["hatchling>=1.27"]
+build-backend = "hatchling.build"
+
+[project]
+name = "hindsight-ext-supabase-tenant"
+version = "0.1.0"
+description = "Supabase Auth tenant extension for Hindsight — per-user schema isolation from Supabase JWTs"
+license = "MIT"
+readme = "README.md"
+requires-python = ">=3.11"
+keywords = ["hindsight", "extension", "supabase", "multi-tenant", "auth"]
+classifiers = [
+    "Development Status :: 4 - Beta",
+    "Intended Audience :: Developers",
+    "License :: OSI Approved :: MIT License",
+    "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
+]
+
+# hindsight-api-slim is deliberately NOT a runtime dependency: the server is the
+# host process that imports this extension, not something the extension installs.
+# Declaring it would let `pip install` of an extension silently upgrade or
+# downgrade the server it is being added to. See ../README.md ("Depend on the
+# server only for development").
+dependencies = [
+    "PyJWT[crypto]>=2.12.0",  # JWKS/RS256 verification; >=2.12 accepts unknown crit header extensions
+    "httpx>=0.27.0",
+]
+
+[project.optional-dependencies]
+dev = [
+    "hindsight-api-slim>=0.9.2",
+    "pytest>=7.0.0",
+    "pytest-asyncio>=0.21.0",
+]
+
+[project.urls]
+Homepage = "https://github.com/vectorize-io/hindsight"
+Documentation = "https://github.com/vectorize-io/hindsight/tree/main/hindsight-extensions/supabase-tenant"
+Repository = "https://github.com/vectorize-io/hindsight"
+
+[tool.uv.sources]
+# Development only — never part of the published wheel's metadata. Tests run
+# against the server checked out beside this package.
+hindsight-api-slim = { path = "../../hindsight-api-slim", editable = true }
+
+[tool.hatch.build.targets.wheel]
+packages = ["hindsight_ext_supabase_tenant"]
+
+[tool.pytest.ini_options]
+testpaths = ["tests"]
+asyncio_mode = "auto"
+asyncio_default_fixture_loop_scope = "function"
