+++ skills/hindsight-docs/references/developer/extensions.md
@@ -19,18 +19,13 @@ HINDSIGHT_API_TENANT_EXTENSION=hindsight_api.extensions.builtin.tenant:ApiKeyTen
 HINDSIGHT_API_TENANT_API_KEY=your-secret-key
 ```
 
-**Built-in: SupabaseTenantExtension**
+**Separately packaged: SupabaseTenantExtension**
 
-Validates [Supabase](https://supabase.com) JWTs and provides multi-tenant memory isolation. Each authenticated user gets their own PostgreSQL schema (`{prefix}_{user_id}`), ensuring complete data separation. Performs local JWT verification using JWKS for optimal performance (no network call per request).
+Validates [Supabase](https://supabase.com) JWTs and gives each authenticated user their own PostgreSQL schema. It ships as its own package, `hindsight-ext-supabase-tenant` — see [`hindsight-extensions/supabase-tenant`](https://github.com/vectorize-io/hindsight/tree/main/hindsight-extensions/supabase-tenant) for configuration and installation.
 
-```bash
-HINDSIGHT_API_TENANT_EXTENSION=hindsight_api.extensions.builtin.supabase_tenant:SupabaseTenantExtension
-HINDSIGHT_API_TENANT_SUPABASE_URL=https://your-project.supabase.co
-# Optional - only needed for legacy HS256 projects or health check
-HINDSIGHT_API_TENANT_SUPABASE_SERVICE_KEY=your-service-role-key
-```
-
-See the [source code](https://github.com/vectorize-io/hindsight/blob/main/hindsight-api-slim/hindsight_api/extensions/builtin/supabase_tenant.py) for complete configuration options and implementation details.
+:::note Moved in 0.10
+Up to 0.9.2 this extension was built in, at `hindsight_api.extensions.builtin.supabase_tenant`. Installs still using that path now fail at startup with migration instructions: `pip install hindsight-ext-supabase-tenant` and set `HINDSIGHT_API_TENANT_EXTENSION=hindsight_ext_supabase_tenant:SupabaseTenantExtension`. All `HINDSIGHT_API_TENANT_*` settings and the schema naming are unchanged.
+:::
 
 For other multi-tenant setups with separate schemas per tenant (e.g., custom JWT-based auth), implement a custom `TenantExtension`.
 
@@ -314,28 +309,19 @@ class MyMCPExtension(MCPExtension):
 
 ### With Docker
 
-Mount your extension package as a volume and set the environment variable:
-
-```yaml
-# docker-compose.yml
-services:
-  hindsight-api:
-    image: vectorize/hindsight-api:latest
-    volumes:
-      - ./my_extensions:/app/my_extensions
-    environment:
-      - HINDSIGHT_API_TENANT_EXTENSION=my_extensions.auth:JwtTenantExtension
-      - HINDSIGHT_API_TENANT_JWT_SECRET=${JWT_SECRET}
-      - PYTHONPATH=/app
-```
-
-Or build a custom image with your extensions:
+Extensions are not bundled in the image, so build one that adds yours to the environment the server runs from. Install into the image's virtualenv explicitly — it was created by `uv sync` and ships no `pip` of its own:
 
 ```dockerfile
-FROM vectorize/hindsight-api:latest
-COPY my_extensions /app/my_extensions
-ENV PYTHONPATH=/app
+FROM ghcr.io/vectorize-io/hindsight:latest
+
+RUN uv pip install --python /app/api/.venv/bin/python --no-cache \
+      hindsight-ext-supabase-tenant \
+ && /app/api/.venv/bin/python -c "import hindsight_ext_supabase_tenant"
 ```
+
+Then point the service at that image and pass the extension's variables as environment. Give the API and worker containers the same extension configuration — the worker uses the tenant extension to enumerate schemas for background consolidation.
+
+See the [extensions registry README](https://github.com/vectorize-io/hindsight/blob/main/hindsight-extensions/README.md#docker-packaging) for the full recipe, including unpublished extensions and dependency pinning.
 
 ### Bare Metal
 
@@ -370,4 +356,6 @@ Custom extensions that solve common use cases are welcome contributions to the H
 - Metrics exporters (Datadog, New Relic, etc.)
 - Custom HTTP endpoints for specific platforms
 
-Consider contributing it to the `hindsight_api.extensions.builtin` package. Open an issue or pull request on [GitHub](https://github.com/vectorize-io/hindsight) to discuss your extension.
+Add it to the [extensions registry](https://github.com/vectorize-io/hindsight/blob/main/hindsight-extensions/README.md) — either as a package under `hindsight-extensions/`, or as a registry entry linking to your own repository. That README covers the package layout, the development and publishing workflow, and Docker packaging.
+
+Extensions live outside the server so that installing Hindsight does not pull in a vendor's client library, and so an extension can be released without a Hindsight release. Only extensions that add no dependencies and are useful to any deployment (`ApiKeyTenantExtension`, `MemoryDefenseRegexExtension`) stay in `hindsight_api.extensions.builtin`.
